feat: Refactor typography
