+++ src/AlgoliaSearch/algoliaSearch.stories.mdx
@@ -6,7 +6,7 @@ import {
   ArgsTable,
 } from "@storybook/addon-docs/blocks";
 import AlgoliaSearch from "./index";
-import Text from "../Typography/Text";
+import Text from "../Text";
 import Content from "../Content";
 
 <Meta title="AlgoliaSearch" component={AlgoliaSearch} />
