+++ src/AlgoliaSearch/elements/ResultsList/SectionTitle/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
-import Text from "../../../../Typography/Text";
+import Text from "../../../../Text";
 
 const SectionTitle = ({ title }) => (
   <div className="py-2 mb-2 top-0 bg-white border-b border-border">
