+++ src/Sidesheet/sidesheet.stories.mdx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
 import Sidesheet from ".";
 import Button from "../Button";
-import Text from "../Typography/Text";
+import Text from "../Text";
 
 <Meta title="Sidesheet" component={Sidesheet} />
