+++ src/Table/index.js
@@ -3,7 +3,7 @@
 import { jsx } from "@emotion/react";
 import React from "react";
 import PropTypes from "prop-types";
-import Text from "../Typography/Text";
+import Text from "../Text";
 import tw from "twin.macro";
 
 const Table = ({ children, ...props }) => (
