+++ src/Tabs/tabs.stories.mdx
@@ -1,6 +1,5 @@
 import { Meta, Story, Preview } from "@storybook/addon-docs/blocks";
 import { Tabs, TabsList, TabsTab, TabsPanel } from "./index";
-import Text from "../Typography/Text";
 
 <Meta title="Tabs" component={Tabs} />
