+++ src/Text.js
