+++ src/Typography/index.js
@@ -1,7 +1,9 @@
-import Text from "./Text";
-import Heading from "./Heading";
-import UnorderedList from "./UnorderedList";
-import ListItem from "./ListItem";
+import Text from "../Text";
+import Heading from "../Heading";
+import UnorderedList from "../UnorderedList";
+import ListItem from "../ListItem";
+
+// These are present for backward compatibility
 
 export {
     Text,
