+++ src/Typography/typography.stories.mdx
@@ -1,8 +1,8 @@
 import { Meta, Story, Preview, ArgsTable } from "@storybook/addon-docs/blocks";
-import Text from "./Text";
-import Heading from "./Heading";
-import UnorderedList from "./UnorderedList";
-import ListItem from "./ListItem";
+import Text from "../Text";
+import Heading from "../Heading";
+import UnorderedList from "../UnorderedList";
+import ListItem from "../ListItem";
 
 <Meta title="Typography" component={Heading} />
